create a separate avgRating resolver
