+++ server/src/avgRating.graphql
@@ -1,0 +1,8 @@
+type AverageRatingsPayload {
+  title: String!
+  content: String!
+}
+
+extend type Query {
+  avgRating: [AverageRatingsPayload!]!
+}
