--- server/src/lockerRatings.graphql
@@ -1,10 +1,0 @@
-subscription {
-  Review(filter: {
-    mutation_in: [CREATED, UPDATED]
-  }) {
-    node {
-      locker
-      rating
-    }
-  }
-}
